forge/pyproject.toml
[project]
name = "forge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datasets>=4.1.1",
    "peft>=0.17.1",
    "pillow>=11.3.0",
    "torch>=2.8.0",
    "transformers>=4.56.2",
    "wandb>=0.22.0",
]
